#25 feat: guestUserAlbum 메서드 userAlbumMapper로 이동 및 앨범 생성 시 hostUserAlbum도 저장
