+++ src/main/java/com/cheeeese/album/infrastructure/mapper/AlbumMapper.java
@@ -1,8 +1,6 @@
 package com.cheeeese.album.infrastructure.mapper;
 
 import com.cheeeese.album.domain.Album;
-import com.cheeeese.album.domain.UserAlbum;
-import com.cheeeese.album.domain.type.Role;
 import com.cheeeese.album.dto.response.AlbumCreationResponse;
 import com.cheeeese.album.dto.response.AlbumEnterResponse;
 import com.cheeeese.album.dto.response.AlbumEnterResponse.AlbumHostInfo;
@@ -106,17 +104,6 @@ public class AlbumMapper {
         return AlbumHostInfo.builder()
                 .hostName(host.getName())
                 .hostProfileImage(host.getProfileImage())
-                .build();
-    }
-
-    /**
-     * User와 Album 엔티티를 기반으로 GUEST 역할의 UserAlbum 엔티티를 생성합니다.
-     */
-    public static UserAlbum toGuestUserAlbum(User user, Album album) {
-        return UserAlbum.builder()
-                .userId(user.getId())
-                .albumId(album.getId())
-                .role(Role.GUEST)
                 .build();
     }
