+++ src/main/java/com/cheeeese/album/infrastructure/mapper/UserAlbumMapper.java
@@ -1,7 +1,9 @@
 package com.cheeeese.album.infrastructure.mapper;
 
+import com.cheeeese.album.domain.Album;
 import com.cheeeese.album.domain.UserAlbum;
 import com.cheeeese.album.domain.type.Role;
+import com.cheeeese.user.domain.User;
 
 public class UserAlbumMapper {
 
@@ -10,6 +12,17 @@ public class UserAlbumMapper {
                 .userId(userId)
                 .albumId(albumId)
                 .role(role)
+                .build();
+    }
+
+    /**
+     * User와 Album 엔티티를 기반으로 GUEST 역할의 UserAlbum 엔티티를 생성합니다.
+     */
+    public static UserAlbum toGuestUserAlbum(User user, Album album) {
+        return UserAlbum.builder()
+                .userId(user.getId())
+                .albumId(album.getId())
+                .role(Role.GUEST)
                 .build();
     }
 }
